requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
zipp==3.11.0
